Data.Field

Description

A Field is a Ring in which all nonzero elements have a multiplicative inverse.

Field typeclass

class (Euclidean a, Ring a) => Field a

Field represents a field, a ring with a multiplicative inverse for any non-zero element.

divide :: Field a => a -> a -> a infixl 7

Divide two elements of a Field. For any Fractional type, this is the same as (/).

x divide y = x times recip y

fromRational :: Field a => Rational -> a

Convert from rational to field.

When {-# LANGUAGE RebindableSyntax #-} is enabled, this function is used for desugaring rational literals (like, 2.37). This may be used to facilitate transition from Fractional to Field, because less casts are now required.

recip :: Field a => a -> a

Invert an element of a Field. For any Fractional type, this is the same as recip.

recip x times x = one

(/) :: Field a => a -> a -> a infixl 7

Infix shorthand for divide.