Data.Semiring

Description

A class for semirings (types with two binary operations, one commutative and one associative, and two respective identities), with various general-purpose instances.

Semiring typeclass

class Semiring a where

The class of semirings (types with two binary operations and two respective identities). One can think of a semiring as two monoids of the same underlying type, with the first being commutative. In the documentation, you will often see the first monoid being referred to as additive, and the second monoid being referred to as multiplicative, a typical convention when talking about semirings.

For any type R with a Num instance, the additive monoid is (R, +, 0) and the multiplicative monoid is (R, *, 1).

For Bool, the additive monoid is (Bool, ||, False) and the multiplicative monoid is (Bool, &&, True).

Instances should satisfy the following laws:

additive left identity

zero + x = x

additive right identity

x + zero = x

additive associativity

x + (y + z) = (x + y) + z

additive commutativity

x + y = y + x

multiplicative left identity

one * x = x

multiplicative right identity

x * one = x

multiplicative associativity

x * (y * z) = (x * y) * z

left-distributivity of * over +

x * (y + z) = (x * y) + (x * z)

right-distributivity of * over +

(x + y) * z = (x * z) + (y * z)

annihilation

zero * x = x * zero = zero

Minimal complete definition

plus, times, (zero, one | fromNatural)

Methods

plus infixl 6

Arguments

:: a

-> a

-> a
Commutative Operation

zero

Arguments

:: a
Commutative Unit

times infixl 7

Arguments

:: a

-> a

-> a
Associative Operation

one

Arguments

:: a
Associative Unit

fromNatural

Arguments

:: Natural

-> a
Homomorphism of additive semigroups

(+) :: Semiring a => a -> a -> a infixl 6

Infix shorthand for plus.

(*) :: Semiring a => a -> a -> a infixl 7

Infix shorthand for times.

(^) :: (Semiring a, Integral b) => a -> b -> a infixr 8

Raise a number to a non-negative integral power. If the power is negative, this will call error.

foldMapP :: (Foldable t, Semiring s) => (a -> s) -> t a -> s

Map each element of the structure to a semiring, and combine the results using plus.

foldMapT :: (Foldable t, Semiring s) => (a -> s) -> t a -> s

Map each element of the structure to a semiring, and combine the results using times.

sum :: (Foldable t, Semiring a) => t a -> a

The sum function computes the additive sum of the elements in a structure. This function is lazy. For a strict version, see sum'.

product :: (Foldable t, Semiring a) => t a -> a

The product function computes the product of the elements in a structure. This function is lazy. for a strict version, see product'.

sum' :: (Foldable t, Semiring a) => t a -> a

The sum' function computes the additive sum of the elements in a structure. This function is strict. For a lazy version, see sum.

product' :: (Foldable t, Semiring a) => t a -> a

The product' function computes the additive sum of the elements in a structure. This function is strict. For a lazy version, see product.

isZero :: (Eq a, Semiring a) => a -> Bool

Is the value zero?

isOne :: (Eq a, Semiring a) => a -> Bool

Is the value one?

Types

newtype Add a

Monoid under plus. Analogous to Sum, but uses the Semiring constraint rather than Num.

Constructors

Add

Fields

getAdd :: a

newtype Mul a

Monoid under times. Analogous to Product, but uses the Semiring constraint rather than Num.

Constructors

Mul

Fields

getMul :: a

newtype WrappedNum a

Provide Semiring and Ring for an arbitrary Num. It is useful with GHC 8.6+'s DerivingVia extension.

Constructors

WrapNum

Fields

unwrapNum :: a

newtype Mod2

Mod2 represents the integers mod 2.

It is useful in the computing of Zhegalkin polynomials.

Constructors

Mod2

Fields

getMod2 :: Bool

newtype IntSetOf a

Wrapper to mimic Set (Sum Int), Set (Product Int), etc., while having a more efficient underlying representation.

Constructors

IntSetOf

Fields

getIntSet :: IntSet

newtype IntMapOf k v

Wrapper to mimic Map (Sum Int) v, Map (Product Int) v, etc., while having a more efficient underlying representation.

Constructors

IntMapOf

Fields

getIntMap :: IntMap v

Ring typeclass

class Semiring a => Ring a where

The class of semirings with an additive inverse.

negate a + a = zero

Methods

negate :: a -> a

fromInteger :: Ring a => Integer -> a

Convert from integer to ring.

When {-# LANGUAGE RebindableSyntax #-} is enabled, this function is used for desugaring integer literals. This may be used to facilitate transition from Num to Ring: no need to replace 0 and 1 with one and zero or to cast numeric literals.

fromIntegral :: (Integral a, Ring b) => a -> b

Convert from integral to ring.

minus :: Ring a => a -> a -> a infixl 6

Subtract two Ring values. For any type R with a Num instance, this is the same as (-).

x minus y = x + negate y

(-) :: Ring a => a -> a -> a infixl 6

Infix shorthand for minus.