Data.Star

Description

A class for *-semirings (pron. "star-semirings").

Documentation

class Semiring a => Star a where

A Star semiring adds one operation, star to a Semiring, such that it follows the law:

star x = one + x * star x = one + star x * x

Another operation, aplus, can be defined in terms of star:

aplus x = x * star x

Minimal complete definition

star | aplus

Methods

star :: a -> a

aplus :: a -> a