Data.Ring.Ordered

Description

An 'ordered ring' is a ring with a total order.

Mathematical pedantry note

Many (if not most) of the instances of the OrderedRing type class are not truly ordered rings in the mathematical sense, as the axioms imply that the underlying set is either a singleton or infinite. Thus, the additional properties of ordered rings do not, in general, hold.

We indicate those instances that are 'truly' or 'mathematically' ordered rings in their documentation.

Helper types

newtype Modular a

A wrapper to indicate the type is being treated as a modular arithmetic system whose modulus is the type's cardinality.

While we cannot guarantee that infinite types won't be wrapped by this, we only provide instances of the relevant type classes for those types we are certain are finite.

Since: 0.7

Constructors

Modular

Fields

getModular :: a

Ordered ring type class

class (Ring a, Ord a) => OrderedRing a where

The class of rings which also have a total order.

Instance should satisfy the following laws:

abs zero = zero

abs x = abs (negate x)

x - abs x = zero

signum zero = zero

If x > zero, then signum x = one

If x < zero, then signum x = negate one

Since: 0.7

Methods

abs :: a -> a

Compute the absolute value.

signum :: a -> a

Determine the 'sign' of a value.