Data.Semiring.Directed

Description

A "directed semiring" refers to the semiring composed of the union of upwards directed sets as multiplication, and intersection of downwards directed sets as addition.

Directed semirings

newtype Directed

Wrapper for the semiring of upwards and downwards directed sets.

For the individual join/meet monoids associated with either algebra, see Max Ordering, and Min Ordering@.

Constructors

Directed

Fields

getDirected :: Ordering

Since: 0.7